[posix] move SMB implementation to platform folder
