+++ xbmc/platform/posix/filesystem/SMBDirectory.cpp
