+++ xbmc/platform/posix/filesystem/SMBFile.cpp
