+++ xbmc/platform/posix/filesystem/SMBFile.h
@@ -27,7 +27,7 @@
 //////////////////////////////////////////////////////////////////////
 
 
-#include "IFile.h"
+#include "filesystem/IFile.h"
 #include "URL.h"
 #include "threads/CriticalSection.h"
